Split resolver.rs into a module
