+++ crates/puffin-resolver/src/lib.rs
@@ -5,9 +5,7 @@ pub use prerelease_mode::PreReleaseMode;
 pub use resolution::ResolutionGraph;
 pub use resolution_mode::ResolutionMode;
 pub use resolution_options::ResolutionOptions;
-pub use resolver::{
-    BuildId, DefaultResolverProvider, Reporter as ResolverReporter, Resolver, ResolverProvider,
-};
+pub use resolver::{BuildId, Reporter as ResolverReporter, Resolver, ResolverProvider};
 
 mod candidate_selector;
 mod error;
